A small tool was added that provides a simple GUI for viewing streams saved in vis-brain, re-playing these streams, their save into PNG and MAT format, saving any selected frame into BIN, PNG or MAT format.
F.A.Q.
(1) Why Matlab/Octave support is not implemented in vis-brain?
This is not assumed that Matlab or Octave is installed on the server. No Matlab means no libraries needed to be attached to vis-brain. Vis-brain with support of .mat will fail to start on machines where Matlab/Octave is not pre-installed

(2) Why .avi streams are not realized?
Video rendering requires Graphic Processing Unit (GPU) that is not assumed to be installed on the server.